Testing functionality working.
