+++ .gitignore
@@ -10,7 +10,6 @@ db/
 
 # Tests
 !/tests/*.pdf
-test_cases.json
 
 # Sphinx documentation
 docs/build/
